vps: compatibility change
fifth tuple element for addExternalChild is False by default, but
versions which do not know the fifth parameter won't show it in the
externals list.
so by omitting this explicit "False" it is still not shown in the webgui
in the most recent version but visible in the externals list of previous
ones.